+++ src/Prelude/Str.php
@@ -43,3 +43,39 @@ function lines(string $input): array
         return trim($row);
     }, preg_split('/\n/', $input));
 }
+
+/**
+ * Checks if a string starts with another string.
+ *
+ * @param string $haystack
+ * @param string $needle
+ * @return bool
+ */
+function starts_with(string $haystack, string $needle): bool
+{
+    return mb_strpos($haystack, $needle) === 0;
+}
+
+/**
+ * Checks if a string ends by another string.
+ *
+ * @param string $haystack
+ * @param string $needle
+ * @return bool
+ */
+function ends_with(string $haystack, string $needle): bool
+{
+    return starts_with(strrev($haystack), strrev($needle));
+}
+
+/**
+ * Checks if a string contains another string.
+ *
+ * @param string $haystack
+ * @param string $needle
+ * @return bool
+ */
+function contains(string $haystack, string $needle): bool
+{
+    return mb_strpos($haystack, $needle) > -1;
+}