+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Better error message for `array_get`'s
 - Add `array_get_arr` function
 - PHPUnit 9
+- Add `starts_with`, `ends_with` and `contains`
 
 # 1.7.2
 - Experimental support for gRPC servers